+++ packages/cybersource-integration/project.json
@@ -1,0 +1,23 @@
+{
+    "root": "packages/cybersource-integration",
+    "sourceRoot": "packages/cybersource-integration/src",
+    "projectType": "library",
+    "targets": {
+        "lint": {
+            "executor": "@nrwl/linter:eslint",
+            "outputs": ["{options.outputFile}"],
+            "options": {
+                "lintFilePatterns": ["packages/cybersource-integration/**/*.ts"]
+            }
+        },
+        "test": {
+            "executor": "@nrwl/jest:jest",
+            "outputs": ["coverage/packages/cybersource-integration"],
+            "options": {
+                "jestConfig": "packages/cybersource-integration/jest.config.js",
+                "passWithNoTests": true
+            }
+        }
+    },
+    "tags": ["scope:integration"]
+}
