+++ tsconfig.base.json
@@ -57,6 +57,9 @@
             "@bigcommerce/checkout-sdk/credit-card-integration": [
                 "packages/credit-card-integration/src/index.ts"
             ],
+            "@bigcommerce/checkout-sdk/cybersource-integration": [
+                "packages/cybersource-integration/src/index.ts"
+            ],
             "@bigcommerce/checkout-sdk/external-integration": [
                 "packages/external-integration/src/index.ts"
             ],
